ok42

+++ exercises/P41_FromWhereToWhere.php
@@ -6,9 +6,9 @@ class P41_FromWhereToWhere
     {
         // Write your program here
         echo "where to?\n";
-        $numberTo = (float)trim(fgets($GLOBALS['STDIN'] ?? STDIN));
+        $numberTo = (int)trim(fgets($GLOBALS['STDIN'] ?? STDIN));
         echo "where from?\n";
-        $numberFrom = (float)trim(fgets($GLOBALS['STDIN'] ?? STDIN));
+        $numberFrom = (int)trim(fgets($GLOBALS['STDIN'] ?? STDIN));
 
         for($i = $numberFrom; $i <= $numberTo; $i++){
                 echo $i."\n";
